+++ flash-core/src/main/java/cn/enilu/flash/service/system/MenuService.java
@@ -226,7 +226,7 @@ public class MenuService  extends BaseService<Menu,Long,MenuRepository> {
         return menuRepository.getMenuIdsByRoleId(roleId);
     }
 
-    public List<String> getResUrlsByRoleId(Integer roleId) {
+    public List<String> getResUrlsByRoleId(Long roleId) {
         return menuRepository.getResUrlsByRoleId(roleId);
     }
 }
